backend/requirements.txt
django
psycopg2-binary
redis
requests
djangorestframework

